client/requirements.txt
grpcio-tools
grpcio
essential-generators
prettytable
humanfriendly